replace some components
