+++ src/components/main/MainComponent.jsx
@@ -1,0 +1,11 @@
+import RoutesComponent from 'routes/Route';
+
+const MainComponent = () => {
+  return (
+    <main>
+      <RoutesComponent />
+    </main>
+  );
+};
+
+export default MainComponent;
